"Added VdjCommand and updated composer dependencies"
This commit introduces the VdjCommand class which is added into the src/Command directory. The purpose of this Symfony command is to manage and operate on virtual DJ (VDJ) tracks, processing them according to certain criteria including point of interest (POI) management and track renaming. In addition, updates are made to the .gitignore file to ignore the JetBrains PhpStorm IDE specific directory (.idea).

In addition, the Doctrine Collections library was added to the project to work with arrays. This library provides classes to work with arrays of data more effectively.

On the other hand, the league/uri package was introduced, along with an update to composer.lock file.  The league/uri is a Simple API to build and manipulate Uniform Resource Identifier.

All these changes are needed to enhance the VDJ track processing feature and improve the way arrays are manipulated within the application.